Yahoo Finance & Moving Averages
Tells you when to buy and sell a stock. You want the red average to be above the white average, followed by the blue average.   The red lines are the shorter term averages and the blue lines are the longer term averages.   The key to a rocket stock is the presence of a white space between the rising short and long term averages.   It is this white space that shows me that the stock is likely to be a rocket. As soon as stock goes RWB, buy it. As soon as it goes BWR, sell it. This program tells you when to do that for a given stock.